backend/requirements.txt
amqp==5.0.2
arabic-reshaper==2.1.1
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.0.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
future==0.18.2
html5lib==1.1
idna==2.10
importlib-metadata==3.3.0
kombu==5.0.2
Pillow==8.0.1
pipfile==0.0.2
prompt-toolkit==3.0.8
PyJWT==1.7.1
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2020.4
reportlab==3.5.57
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.5
zipp==3.4.0
cos-python-sdk-v5==1.9.0
celery==5.0.5
django-celery-beat==2.2.0
django-otp==1.0.2
qrcode==6.1
django-email-verification==0.1.0
django-ical==1.7.1
wechatpy==1.8.15
tencentcloud-sdk-python==3.0.388
django-ckeditor==6.1.0